adding report

+++ app/Owner.php
@@ -52,57 +52,72 @@ class Owner extends Model
 
        static::deleting(function($owner) {
 
-        
-
-
-
 
         $stocks=$owner->stocks;
-        foreach ($stocks as $v) {
-            $v->delete();
-        }
+        
+        if(!empty($stocks)){
 
-        $medicines=$owner->medicines;
-         foreach ($medicines as $v) {
+             foreach ($stocks as $v) {
             $v->delete();
+            }
+
+            $medicines=$owner->medicines;
+             foreach ($medicines as $v) {
+                $v->delete();
+            }
+
         }
+       
         
 
         $treatments=$owner->treatments;
 
-        foreach ($treatments as $v) {
+        if(!empty($treatments)){
+
+              foreach ($treatments as $v) {
             $v->medicines()->detach();
             $v->delete();
+            }
+
+            $appointments=$owner->appointments;
+            foreach ($appointments as $v) {
+                $v->delete();
+            }
+
+            $treatments=$owner->treatments;
+
+            foreach ($treatments as $v) {
+                $v->medicines()->detach();
+                $v->delete();
+            }
+
+            $patients=$owner->patients;
+            foreach ($patients as $v) {
+                $v->delete();
+            }
+
+
         }
 
-        $appointments=$owner->appointments;
-        foreach ($appointments as $v) {
-            $v->delete();
-        }
-
-        $treatments=$owner->treatments;
-
-        foreach ($treatments as $v) {
-            $v->medicines()->detach();
-            $v->delete();
-        }
-
-        $patients=$owner->patients;
-        foreach ($patients as $v) {
-            $v->delete();
-        }
-
+      
         $doctors=$owner->doctors;
-        foreach ($doctors as $v) {
-            $v->user->removeRole('Doctor');
-            $v->delete();
+        if(!empty($doctors)){
+            foreach ($doctors as $v) {
+                $v->user->removeRole('Doctor');
+                $v->delete();
+            }
         }
+        
 
         $receptions=$owner->receptions;
-        foreach ($receptions as $v) {
-            $v->user->removeRole('Reception');
-            $v->delete();
+        if(!empty($receptions)){
+                foreach ($receptions as $v) {
+                $v->user->removeRole('Reception');
+                $v->delete();
+            }
         }
+        
+
         $owner->user->removeRole('Admin');
         $owner->user->delete();
